+++ yt_dlp/extractor/extractors.py
@@ -308,6 +308,7 @@ from .democracynow import DemocracynowIE
 from .dfb import DFBIE
 from .dhm import DHMIE
 from .digg import DiggIE
+from .discoveryplusindia import DiscoveryPlusIndiaIE
 from .dotsub import DotsubIE
 from .douyutv import (
     DouyuShowIE,
@@ -316,7 +317,6 @@ from .douyutv import (
 from .dplay import (
     DPlayIE,
     DiscoveryPlusIE,
-    DiscoveryPlusIndiaIE,
     HGTVDeIE,
 )
 from .dreisat import DreiSatIE
